split dataclass ui, path inheritance
