+++ omc3_gui/segment_by_segment/measurement_model.py
@@ -23,7 +23,7 @@ from omc3.optics_measurements.constants import (
 from tfs.reader import read_headers
 
 from omc3_gui.ui_components.dataclass_ui import choices_validator as choices
-from omc3_gui.ui_components.dataclass_ui import metafield
+from omc3_gui.ui_components.dataclass_ui import metafield, DirectoryPath, FilePath
 
 if TYPE_CHECKING:
     from omc3_gui.segment_by_segment.segment_model import SegmentDataModel
@@ -43,14 +43,14 @@ class OpticsMeasurement:
     which can then be passed on to the segment-by-segment.
     The :func:`omc3_gui.utils.dataclass_ui.metafield` is used to provide hints about the fields for the GUI.
     """
-    measurement_dir: Path = metafield("Optics Measurement", "Path to the optics-measurement folder")
-    model_dir: Path =       metafield("Model",              "Path to the model folder",        default=None)
-    accel: str =            metafield("Accelerator",        "Name of the accelerator",         default=None)
-    output_dir: Path =      metafield("Output",             "Path to the sbs-output folder",   default=None) 
-    corrections: Path =     metafield("Corrections",        "Path to the corrections file",    default=None)
-    year: str =             metafield("Year",               "Year of the measurement (model)", default=None)
-    ring: int =             metafield("Ring",               "Ring of the accelerator",         default=None, validate=choices(1, 2, 3, 4))
-    beam: int =             metafield("Beam",               "Beam of the accelerator",         default=None, validate=choices(1, 2)) 
+    measurement_dir: DirectoryPath = metafield("Optics Measurement", "Path to the optics-measurement folder")
+    model_dir: DirectoryPath =       metafield("Model",              "Path to the model folder",        default=None)
+    accel: str =                     metafield("Accelerator",        "Name of the accelerator",         default=None)
+    output_dir: DirectoryPath =      metafield("Output",             "Path to the sbs-output folder",   default=None) 
+    corrections: FilePath =          metafield("Corrections",        "Path to the corrections file",    default=None)
+    year: str =                      metafield("Year",               "Year of the measurement (model)", default=None)
+    ring: int =                      metafield("Ring",               "Ring of the accelerator",         default=None, validate=choices(1, 2, 3, 4))
+    beam: int =                      metafield("Beam",               "Beam of the accelerator",         default=None, validate=choices(1, 2)) 
     # List of segments. Using a list here, so the name and start/end can be changed
     # without having to modify anything here.
     _segments: list[SegmentDataModel] = field(default_factory=list)
@@ -168,7 +168,7 @@ class OpticsMeasurement:
 
     # Builder ------------------------------------------------------------------
     @classmethod
-    def from_path(cls, path: Path) -> "OpticsMeasurement":
+    def from_path(cls, path: Path) -> OpticsMeasurement:
         """ Creates an OpticsMeasurement from a folder, by trying 
         to parse information from the data in the folder.
