+++ omc3_gui/ui_components/dataclass_ui.py
@@ -26,606 +26,11 @@ LOGGER = logging.getLogger(__name__)
 
 
 # Helper for the dataclass definitions -----------------------------------------
-@dataclass(slots=True)
-class MetaData:
-    """ Metadata for a dataclass-field. """
-    label: str | None = None
-    comment: str | None = None
-    validate: Callable | None = None  # needs to return a truthy value if valid, else falsy or raise
 
-    def __getitem__(self, key):
-        return getattr(self, key)
-
-    def get(self, key, default=None):
-        return getattr(self, key, default)
-
-    
-def metafield(
-    label: str, 
-    comment: str, 
-    default=MISSING, 
-    validate: Callable | None = None, 
-    ) -> Field:
-    """ Convenience function to create a dataclass-field with metadata. """
-    return field(
-        default=default, 
-        metadata=MetaData(label=label, comment=comment, validate=validate)
-    )
-
-
-def choices_validator(*choices: Any) -> Callable:
-    """ Return a validator that checks if a given value is in the choices. """
-    def validator(value):
-        if value not in choices:
-            raise ValueError(f"Value {value} is not in {choices}.")
-        return True
-    return validator
-
-
-class FilePath(Path):
-    """ Convenience Class to indicate that the Path should lead to a file. """
-    pass
-
-
-class DirectoryPath(Path):
-    """ Convenience Class to indicate that the Path should lead to a directory. """
-    pass
-
-
-
-# DataClass UI Controller -------------------------------------------------------- 
-
-@dataclass(slots=True)
-class FieldUIDef:
-    """ Definition of an FieldUI to be generated by 
-    :func:`omc3_gui.utils.dataclass_ui.dataclass_ui_builder`.
-    This defines how the label and edit-field widgets should look like
-    for a field in the dataclass to be UI'ed.
-    Missing information is parsed in from the dataclass 
-    isetlf if possible, but values given here take precedence. 
-    """
-    name: str                        # name of the field in the dataclass
-    label: str | None = None         # label of the field
-    pytype: type | None = None       # type of the field's data, needs to be instanciable
-    comment: str | None = None       # comment for the field, e.g. used for tooltips
-    editable: bool | None = True     # sets field to be editable
-
-
-@dataclass()
-class FieldUI:
-    """ UI-container for a dataclass field, 
-    as generated by :meth:`omc3_gui.utils.dataclass_ui.DataClassUI.dataclass_ui_builder`
-    and attached to :class:`omc3_gui.utils.dataclass_ui.DataClassUI`.
-    """
-    widget: QtWidgets.QWidget  # edit-widget to edit field value
-    label: QtWidgets.QLabel    # label-widget of the field
-    get_value: Callable        # getter for the widget value, returns the value as appropriate type for the dataclass
-    set_value: Callable        # setter for the widget value
-    text_color: str | None = colors.TEXT_DARK  # default text-color for both widget and label
-    modified: bool = False     # flag indicating if the widget-content has been modified by the user
-
-    def __post_init__(self):
-        """ Connects the widget to the label and sets the text-color. """
-        for main_attribute in ["text", "value", "state"]:
-            try:
-                signal = getattr(self.widget, f"{main_attribute}Changed")
-            except AttributeError:
-                pass
-            else:
-                signal.connect(self.has_changed)
-                break
-        self.widget.setStyleSheet(f"color: {self.text_color};")
-        self.label.setStyleSheet(f"QLabel {{color: {self.text_color}}};") 
-
-    def has_changed(self):
-        """ Triggered when the widget has been modified.
-        Sets then the modified flag and changes label font.
-         """
-        self.modified = True
-        font = self.label.font()
-        font.setItalic(True)
-        self.label.setFont(font)
-    
-    def reset(self):
-        """ Reset the label font to normal and clear the modified flag. """
-        self.modified = False
-        font = self.label.font()
-        font.setItalic(False)
-        self.label.setFont(font)
-
-
-class DataClassUI:
-    """ Controller for the UI representation of a dataclass. 
-    It contains a grid-layout that can be added to any QWidget/QLayout.   
-    """
-
-    def __init__(self, field_definitions: Sequence[FieldUIDef], dclass: type | object, layout: QtWidgets.QLayout | None = None):
-        """ Builds a DataClassUI from a list of field definitions.
-        NOTE: `dclass` is not automatically attached to the resulting class, unless it is an already an instance.
-        This function works with classes and instances, but the attached object needs to be the instance.
-        
-        Args:
-            field_def (Sequence[FieldUIDef | str]): list of field definitions
-            dclass (type | object): DataClass type or instance.
-            layout (QtWidgets.QLayout | None): Layout to add the UI to. If None, a new grid-layout is created.
-
-        Returns:
-            DataClassUI: An instances with a grid-layout containing edit-widgets and labels.
-        """
-        if layout is None:
-            layout = QtWidgets.QGridLayout()
-        
-        self.model: object = None if isinstance(dclass, type) else dclass  # dataclass instance 
-        self.layout: QtWidgets.QGridLayout = layout  # final layout of the UI for dataclass
-        self.fields: dict[str, FieldUI] = {} # stored field UI-elements
-
-        field_instances: dict[str, Field] = {field.name: field for field in fields(dclass)}
-        field_types = get_dataclass_types(
-            dclass, 
-            [
-                field.name for field in field_definitions 
-                if field is not None and not isinstance(field, str) and field.name in field_instances.keys()
-             ]
-        )
-        # 
-        
-        for idx_row, field_def in enumerate(field_definitions):
-            if field_def is None:  # Separator ---
-                layout.addWidget(HorizontalSeparator(), idx_row, 0, 1, 3)
-                continue
-
-            if isinstance(field_def, str):  # Label only ---
-                layout.addWidget(QtWidgets.QLabel(field_def), idx_row, 0)
-                continue
-            
-            if field_def.name not in field_instances:
-                raise ValueError(f"Field {field_def.name} not found in dataclass {dclass}")
-
-            field_inst = field_instances[field_def.name]
-            
-            # Label ---
-            qlabel = QtWidgets.QLabel(field_def.label or field_inst.metadata.get("label", field_def.name))
-            qlabel.setToolTip(field_def.comment or field_inst.metadata.get("comment", ""))
-            layout.addWidget(qlabel, idx_row, 0)
-            
-            # User input ---
-            # If field_def.type is not given, use evaluate from dataclass. 
-            # Check __args__ in case of Union/Optional and use first one.
-            # The type needs to be instanciable!
-            eval_type = field_types[field_def.name]  # evaluated type, might be Optional/Union etc with __args__
-            field_type = field_def.pytype or getattr(eval_type, "__args__", [eval_type])[0]  
-
-            widget = TYPE_TO_WIDGET_MAP.get(field_type, QtWidgets.QLineEdit)()
-
-            try:
-                widget.setReadOnly(not field_def.editable)
-            except AttributeError:
-                widget.setEnabled(field_def.editable)
-
-            get_value, set_value = build_getter_setter(widget, field_type)
-            self.fields[field_def.name] = FieldUI(
-                widget=widget, 
-                label=qlabel,
-                set_value=set_value,
-                get_value=get_value,
-                text_color=colors.TEXT_DARK if field_def.editable else colors.GREYED_OUT_TEXT_DARK
-            )
-            
-            if not issubclass(field_type, Path) or not field_def.editable:
-                layout.addWidget(widget, idx_row, 1, 1, 2)
-            else:
-                layout.addWidget(widget, idx_row, 1)
-                # Add Path selection button ---
-                button = QtWidgets.QPushButton("...")
-                button.setFixedWidth(30)
-                
-                if field_type is FilePath:
-                    dialog = file_dialogs.OpenFileDialog
-
-                elif field_type is DirectoryPath:
-                    dialog = file_dialogs.OpenDirectoryDialog 
-
-                else:
-                    dialog = file_dialogs.OpenAnySingleDialog
-
-                button.clicked.connect(partial(
-                    run_dialog, 
-                    dialog=dialog,
-                    get_value=get_value,
-                    set_value=set_value))
-                layout.addWidget(button, idx_row, 2)
-
-        if self.model is not None:
-            self.update_ui()  # fill with values from the instance
-
-    def reset_labels(self):
-        """ Resets all labels to indicate that the field shows the currently set value in the dataclass."""
-        for name in self.fields.keys():
-            self.fields[name].reset()
-    
-    def update_widget_from_model(self, name: str):
-        """ Updates the edit-widget of the given field from the dataclass values. 
-        
-        Args:
-            name (str): name of the field in the dataclass
-        """
-        if self.model is None:
-            raise ValueError("No dataclass instance attached to the UI.")
-
-        value = getattr(self.model, name)
-        if value is not None:
-            self.fields[name].set_value(value)
-    
-    def update_model_from_widget(self, name: str):
-        """ Updates the dataclass value of the given field from the edit-widget. 
-        
-        Args:
-            name (str): name of the field in the dataclass
-        """
-        if self.model is None:
-            raise ValueError("No dataclass instance attached to the UI.")
-
-        field: FieldUI = self.fields[name]
-
-        if not field.modified:  # avoid replacing 'None' with widget defaults
-            LOGGER.debug(f"Field {name} was not modified.")
-            return 
-
-        value = field.get_value()
-        setattr(self.model, name, value)
-
-    def update_ui(self):
-        """ Updates all edit-widgets from the dataclass values. """
-        for name in self.fields.keys():
-            self.update_widget_from_model(name)
-
-    def update_model(self):
-        """ Updates all dataclass fields from the current edit-widget values. """
-        for name in self.fields.keys():
-            self.update_model_from_widget(name)
-
-    def validate(self, only_modified: bool = False):
-        """ Checks all edit-widgets for valid choices. 
-        
-        The validation function NEED to return a truthy value, if 
-        the choice is valid. If it is invalid, they can either return a falsy
-        value or raise an exception. In the latter case, this exception will be
-        printed instead of the default message.
-        """
-        if self.model is None:
-            raise ValueError("No dataclass instance attached to the UI.")
-
-        invalid_fields_str = []
-        for name in self.fields.keys():
-            field = self.model.__dataclass_fields__[name]
-            field_ui = self.fields[field.name]
-            if only_modified and not field_ui.modified:
-                continue
-            
-            validate_function = field.metadata.get("validate")
-            if validate_function is not None:
-                value = self.fields[field.name].get_value()
-
-                validation_result = False
-                try: 
-                    validation_result = validate_function(value)
-                except ValueError as e:
-                    invalid_fields_str.append(f"{field_ui.label.text()}: {str(e)}")
-                else: 
-                    if not validation_result:
-                        invalid_fields_str.append(f"{field_ui.label.text()}: {value} is not a valid choice.")
-
-        if invalid_fields_str:
-            full_str = "\n".join(invalid_fields_str)
-            raise ValueError(f'The following fields contain wrong values:\n{full_str}')
-
-
-class DataClassTabbedUI():
-    """ Controller for the UI representation of a dataclass, containing other dataclasses as fields. 
-    It contains a tabbed layout that can be added to any QWidget.   
-    """
-
-    def __init__(self, 
-                dclass: object | type, 
-                layout: QtWidgets.QLayout  | None = None, 
-                widget: QtWidgets.QTabWidget | None = None
-        ):
-        """ Builds a DataClassTabbedUI from a dataclass, using DataClassUIs as widgets in the tabs.
-        NOTE: As with the DataClassUI, the dataclass (`dclass`) is not automatically
-              attached to the resulting class, 
-        
-        Args:
-            dclass (object | type): dataclass instance or class. 
-                                    Needs to contain dataclasses as fields, which will converted to tabs.
-            layout (QtWidgets.QLayout | None): Layout to add the UI widget to. 
-                                               If None, a new QVBoxLayout is created.
-            widget (QtWidgets.QTabWidget | None): TabWidget to add the UI tabs to.
-                                                  If None, a new QTabWidget is created.
-                                                  The widget is always added to the layout.
-        """
-        self.layout: QtWidgets.QLayout = layout or QtWidgets.QVBoxLayout()
-        self.widget: QtWidgets.QTabWidget = widget or QtWidgets.QTabWidget()
-        self._dataclass_uis: dict[str, DataClassUI] = {}
-
-        self.layout.addWidget(self.widget)
-
-        fields = [field.name for field in dc.fields(dclass) if field.name[0] != "_"]
-        for field_name in fields:
-            sub_dclass = getattr(dclass, field_name)
-            dataclass_ui = DataClassUI(
-                field_definitions=[FieldUIDef(field.name) for field in dc.fields(sub_dclass) if field.name[0] != "_"],
-                dclass=sub_dclass,
-            )
-            ui_widget = QtWidgets.QWidget()
-            ui_widget.setLayout(dataclass_ui.layout)
-            self.widget.addTab(ui_widget, field_name.capitalize())
-            self._dataclass_uis[field_name] = dataclass_ui
-        
-        self._model: object = None
-        if not isinstance(dclass, type):
-            self.model = dclass  # also updates ui with values
-        
-    @property
-    def model(self):
-        return self._model
-    
-    @model.setter
-    def model(self, model: object):
-        """ Sets the model and updates the dataclass_uis. 
-        This does not add new dataclass_uis to the layout and throws errors 
-        if fields in the model are missing. """
-        self._model = model
-        for name, dataclass_ui in self._dataclass_uis.items():
-            dataclass_ui.model = getattr(model, name)
-            dataclass_ui.update_ui()
-
-    # Function to pass on to dataclass_uis ---
-    def update_ui(self):
-        for dataclass_ui in self._dataclass_uis.values():
-            dataclass_ui.update_ui()
-
-    def update_model(self):
-        for dataclass_ui in self._dataclass_uis.values():
-            dataclass_ui.update_model()
-    
-    def validate(self, only_modified: bool = False):
-        for dataclass_ui in self._dataclass_uis.values():
-            dataclass_ui.validate(only_modified=only_modified)
-    
-    def reset_labels(self):
-        for dataclass_ui in self._dataclass_uis.values():
-            dataclass_ui.reset_labels()
 
 
 # View -------------------------------------------------------------------------
 
-class DataClassInterface(Protocol):
-    """ Protocol for the DataClassDialog. 
-    Observed by DataClassUI and DataClassTabbedUI which can hence be used with
-    the dialog.
-    """
-    layout: QtWidgets.QLayout
-    model: object
-
-    def update_ui(self): ...
-    def update_model(self): ...
-    def validate(self, only_modified: bool = False): ...
-    def reset_labels(self): ...
-
-
-class DataClassDialog(QtWidgets.QDialog):
-    """ Simple dialog window to display the DataClassUI layout. 
-    Adds some convenience functionality like "Ok" and "Cancel" buttons
-    and automatic data-validation on close.
-    """
-
-    WINDOW_TITLE = "Edit DataClass"
-    DEFAULT_SIZE = (800, 600)  # width, height, use -1 for auto
-    
-    def __init__(self, dataclass_ui: DataClassInterface, parent = None):
-        super().__init__(parent)
-        self._button_box: QtWidgets.QDialogButtonBox = None
-        
-        self._dataclass_ui: DataClassInterface = dataclass_ui
-        self._build_gui()
-        self._connect_signals()
-        self._set_size(width=self.DEFAULT_SIZE[0], height=self.DEFAULT_SIZE[1])
-        self.update_ui()
-
-
-    def _set_size(self, width: int = -1, height: int = -1):
-        # Set position to the center of the parent (does not work in WSL for me, jdilly 2023)
-        # parent = self.parent()
-        # if parent is not None:
-        #     parent_geo = parent.geometry()
-        #     parent_pos = parent.mapToGlobal(parent.pos())  # multiscreen support
-        #     if width >= 0:
-        #         x = parent_pos.x() + parent_geo.width() / 2
-        #     else:
-        #         x = parent_pos.x() + (parent_geo.width() - width) / 2
-
-        #     if height >=0 :
-        #         y = parent_pos.y() + parent_geo.height() / 2
-        #     else:
-        #         y = parent_pos.y() + (parent_geo.height() - height) / 2
-        #     self.move(x, y)
-        
-        # Set size
-        self.resize(width, height)
-
-    def _build_gui(self):
-        self.setWindowTitle(self.WINDOW_TITLE)
-        layout = QtWidgets.QVBoxLayout()
-        layout.addLayout(self._dataclass_ui.layout)
-
-        QBtn = QtWidgets.QDialogButtonBox.Ok | QtWidgets.QDialogButtonBox.Cancel
-        self._button_box = QtWidgets.QDialogButtonBox(QBtn)
-        layout.addWidget(self._button_box)
-
-        self.setLayout(layout)
-    
-    def _connect_signals(self):
-        self._button_box.accepted.connect(self.accept)
-        self._button_box.rejected.connect(self.reject)
-
-    def update_ui(self, new_model: object = None):
-        if new_model is not None:
-            self._dataclass_ui.model = new_model 
-        self._dataclass_ui.update_ui()  # triggers changes, so the labels appear in "changed" state
-        self._dataclass_ui.reset_labels()  # so we reset them
-
-    def accept(self):
-        try:
-            self._dataclass_ui.validate(only_modified=True)
-        except ValueError as e:
-            QtWidgets.QMessageBox.critical(self, "Error", str(e))
-            return
-
-        self._dataclass_ui.update_model()
-        super().accept() 
-
-
-class SettingsDialog(DataClassDialog):
-    """ Slight modification of the DataClassDialog to be used for Tabbed-Settings. """
-
-    WINDOW_TITLE = "Settings"
-    DEFAULT_SIZE = (800, -1)
-    
-    def __init__(self, settings: object, parent=None):
-        dataclass_tabbed_ui = DataClassTabbedUI(dclass=settings)       
-        super().__init__(dataclass_ui=dataclass_tabbed_ui, parent=parent)
-
-    @property
-    def settings(self) -> object:
-        return self._dataclass_ui.model
-
-
-# Type-to-Widget Helpers ----------------------------------------------------------------
-
-class QFullIntSpinBox(QtWidgets.QSpinBox):
-    """ Like a QSpinBox, but overwriting default range(0,100) with maximum integer range. """
-
-    def __init__(self, parent: QtWidgets.QWidget | None = None) -> None:
-        super().__init__(parent)
-        self.setRange(-2**31, 2**31 - 1)  # range of signed 32-bit integers 
-
-
-TYPE_TO_WIDGET_MAP = {
-    int: QFullIntSpinBox,  # Maybe just use QLineEdit as well?
-    float: QtWidgets.QLineEdit,
-    str: QtWidgets.QLineEdit,
-    Path: QtWidgets.QLineEdit,
-    bool: QtWidgets.QCheckBox,
-}
-
-
-def build_getter_setter(widget: QtWidgets.QWidget, field_type: type) -> tuple[Callable, Callable]:
-    """ Getter/Setter Factory for widgets. 
-    
-    Args:
-        widget (QtWidgets.QWidget): The widget to get/set the value of.
-        field_type (type): The type of the dataclass field.
-    """
-    if isinstance(widget, QtWidgets.QCheckBox):
-        def get_value() -> bool:
-            return widget.isChecked()
-
-        def set_value(value: bool):
-            widget.setChecked(value)
-    
-    elif isinstance(widget, QtWidgets.QSpinBox):
-        def get_value():
-            return field_type(widget.value())
-
-        def set_value(value):
-            widget.setValue(value)
-        
-    else:   # Any kind of widget should be able to handle strings.      
-        def get_value():
-            return field_type(widget.text())
-
-        def set_value(value):
-            widget.setText(str(value))
-    return get_value, set_value        
-
-
-def run_dialog(dialog: file_dialogs.OpenFilesDialog, get_value: Callable, set_value: Callable):
-    """ Asks the user to select a directory/file, using the `get_value` function 
-    o determine the default directory and passing the result to `set_value` if not `None`.
-
-    Args:
-        dialog (file_dialogs.OpenFilesDialog): The dialog to run.
-        get_value (Callable): The function to get the last value, e.g. from a widget.
-        set_value (Callable): The function to set the value returned from the dialog.
-    """
-    path = dialog(directory=get_value().parent).run_selection_dialog()
-    if path is not None:
-        set_value(path)
-
 
 # Other ------------------------------------------------------------------------
 
-def get_field_inline_comments(dclass: type) -> dict[str, str]:
-    """
-    Returns a dictionary mapping field names to their associated inline code-comments.
-    Has been replaced by the use of the metadata, but I like the function,
-    so I leave it here. (jdilly 2023)
-
-    Parameters:
-        dclass (type): The data class to extract field comments from.
-
-    Returns:
-        Dict[str, str]: A dictionary mapping field names to their associated comments.
-    """
-    matcher = re.compile(r"^(?P<field>[a-zA-Z_]+)\s*:\s*[^#]+#\s*(?P<comment>.*)\s*$")
-    source = inspect.getsource(dclass)
-
-    found_fields = {}
-    for line in source.splitlines()[2:]:  # first two is @dataclass and name
-        line = line.strip()
-        if line.startswith('def '):
-            break
-
-        match = matcher.match(line)
-        if match:
-            found_fields[match.group('field')] = match.group('comment')
-
-    return found_fields
-
-
-def get_dataclass_types(dclass: type | object, names: Sequence[str]):
-    """ 
-    Returns a dictionary mapping field names to their associated types.
-
-    Why this weird way?
-    First of all, we avoided the cyclic imports of type-hints by using 
-    the `from __future__ import annotations`, but whith 
-    this, all type-hints on the dataclasses become `ForwardReferences`, 
-    i.e. strings. So if we get them from the dataclass-fields
-    via `field.pytype` or `field.pytype.__args__`, we cannot instanciate them later on.
-
-    `get_type_hints()` from `typing` solves this problem, but it 
-    evaluates the type-hints of all fields in the dataclass. 
-    But this can causes again import errors, because the classes were not actually imported
-    (this is why there are no cyclic imports)!
-
-    So this function allows using `get_type_hints()` on the dataclass 
-    itself, but only gets the types of the fields that we need, in the hopes 
-    that there is no import problem with those. 
-    
-    If you are here because there is, you need to redesing your dataclass/imports.
-    Sorry.
-
-    Args:
-        dclass (type): The data class to extract field types from.
-        names (Sequence[str]): The names of the fields to extract types from.
-
-    Returns:
-        Dict[str, type]: A dictionary mapping field names to their associated types. 
-    """
-    required_annotations = {
-        name: value for name, value in dict(dclass.__annotations__).items() if name in names
-    }
-    class SafeMock:
-        __annotations__ = required_annotations
-    return get_type_hints(SafeMock)
